requirements.txt
pyyaml
argparse
numpy
torch
torchvision